♻️ Split ISCNForm into field groups
754 lines serving both the wizard and the edit flow. Lifts four contiguous
field groups into app/components/iscn/, leaving ISCNForm as the UForm, the
validation, the dirty-snapshot guards and the file-derived blocks.

formData stays one defineModel passed down rather than decomposed to match the
components: it is a single on-chain payload and hasUnsavedChanges compares it
whole, so splitting the model would mean reassembling it for the dirty check.

Two things moved with their markup rather than staying behind. The AI-suggest
handler goes with its button, since everything it reads and writes is the
classification group's. The prefilled-from-file hint becomes a composable —
four of the groups need it, so the alternative was four copies.

hasSuggested stays with the host: it means the suggested keywords are not saved
yet, which is resetSnapshot's business, not the group's.
